Enhance UI. Revise validation logic.

+++ app/src/main/java/com/vincent/forexledger/fragment/BookDetailFragment.kt
@@ -40,7 +40,7 @@ class BookDetailFragment : Fragment() {
         initToolbar(args.bookName)
 
         btnCreateEntry.setOnClickListener {
-            findNavController().navigate(BookDetailFragmentDirections.toCreateEntry(bookId, book.balance))
+            findNavController().navigate(BookDetailFragmentDirections.toCreateEntry(bookId, book.balance.toFloat()))
         }
         ViewUtils.setInvisible(btnCreateEntry)
